# ISPN-16329 RESP should return only RESP3 response types