--- a/server/resp/src/main/java/org/infinispan/server/resp/RespHandler.java
+++ b/server/resp/src/main/java/org/infinispan/server/resp/RespHandler.java
@@ -8,6 +8,7 @@ import org.infinispan.commons.util.concurrent.CompletionStages;
 import org.infinispan.server.resp.logging.AccessLoggerManager;
 import org.infinispan.server.resp.logging.Log;
 import org.infinispan.server.resp.logging.RespAccessLogger;
+import org.infinispan.server.resp.serialization.Resp3Response;
 import org.infinispan.util.logging.LogFactory;
 
 import io.netty.buffer.ByteBuf;
@@ -182,7 +183,7 @@ public class RespHandler extends ChannelInboundHandlerAdapter {
    @Override
    public void exceptionCaught(ChannelHandlerContext ctx, Throwable cause) {
       log.unexpectedException(cause);
-      ByteBufferUtils.stringToByteBuf("-ERR Server Error Encountered: " + cause.getMessage() + "\\r\\n", requestHandler.allocatorToUse);
+      Resp3Response.error(cause, requestHandler.allocator());
       flushBufferIfNeeded(ctx, false, null);
       ctx.close();
    }