--- a/server/resp/src/test/java/org/infinispan/server/resp/util/ByteBufferUtilsTest.java
+++ b/server/resp/src/test/java/org/infinispan/server/resp/util/ByteBufferUtilsTest.java
@@ -2,7 +2,7 @@ package org.infinispan.server.resp.util;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import org.infinispan.server.resp.ByteBufferUtils;
+import org.infinispan.server.resp.serialization.ByteBufferUtils;
 import org.testng.annotations.Test;
 
 @Test(groups = "unit", testName = "resp.util.ByteBufferUtilsTest")