https://issues.redhat.com/browse/ISPN-16329

A huge diff but it should be +OK to understand. The most important part of the PR is the classes in the `serialization` package. The diff touches all commands to replace the BiConsumer usage with the new serialization. AFAICT, there were fixes to:

- CLIENT TRACKINGINFO
- COMMAND
- HELLO
- MEMORY STATS
- HGETALL
- HINCRBY
- HINCRBYFLOAT
- SDIFF
- SINTER
- ZSCORE
- ZRANK
- ZRANGE
- ZRANDMEMBER
- ZDIFF/ZDIFFSTORE
- ZINTER/ZUNION
- CONFIG GET

Additionally, any command that includes double values in the response. I created a few issues for later, to avoid too many changes: [ISPN-16471](https://issues.redhat.com/browse/ISPN-16471), [ISPN-16460](https://issues.redhat.com/browse/ISPN-16460), and [ISPN-16474](https://issues.redhat.com/browse/ISPN-16474).

I haven't added all the types from the specification. For example, we never use "push" messages or verbatim strings. I have added big numbers, even though we don't use them. We can add the missing pieces when necessary.

To give more context. I removed every "manual" serialization<sup>1</sup>, where we write to a buffer *while* processing a request. Now we have each step in distinct places, the commands only handle a request and after the return, a place to deal with serialization. The classes for serialization are private to the `serialization` package and don't leak outside. We call the appropriate static method of the `Resp3Response` class to interact with serialization. The goal with this was just to split the concerns.

This approach has some downsides. We utilize a registry to add all the specific serializers, and during the runtime, we need to identify which one to use. To list a few of the issues:

* More operations are needed to serialize a response;
* More garbage generated while iterating in the serializers;
* Buffer resizing because we serialize the object recursively without calculating the total size.

These issues mostly affect the operations with aggregate response types (arrays, sets, hashes). Simpler operations that return primitive types should find the appropriate serializer generating less garbage and allocate the buffer with the exact size. But the issues are still there nevertheless.

Honestly, I am not attached to this specific solution. My goal was to separate the serialization from the command handling. If you have other suggestions that don't suffer from the same issues or are more lightweight, I am more than happy to apply them to this PR.

---

1. The `EXEC` and PubSub listeners still need to do some "manual" handling of objects to work. I couldn't remove all the manual pieces for these commands.