@jabolina 2024-08-29
I ran a comparison with redis-benchmark with this branch and main. Running locally (so, there's noise in the system) with a single node. Warming up by running 3x `redis-benchmark -h localhost -p 11222 -n 100000 --threads 12 -d 10000 --user admin -a password -3 -t set,get` and final measurement with 1_000_000 requests:

```bash
redis-benchmark -h localhost -p 11222 -n 1000000 --threads 12 -d 10000 --user admin -a password -3 -t set,get
```

**This branch**

```text
SET:
Summary:
  throughput summary: 76846.23 requests per second
  latency summary (msec):
          avg       min       p50       p95       p99       max
        0.633     0.120     0.599     0.791     1.415    16.463


GET:
Summary:
  throughput summary: 199560.95 requests per second
  latency summary (msec):
          avg       min       p50       p95       p99       max
        0.192     0.016     0.167     0.383     0.727    20.911
```

**main**

```text
SET:
Summary:
  throughput summary: 78382.19 requests per second
  latency summary (msec):
          avg       min       p50       p95       p99       max
        0.626     0.448     0.599     0.775     1.343    11.759
        
        
GET:
Summary:
  throughput summary: 199560.95 requests per second
  latency summary (msec):
          avg       min       p50       p95       p99       max
        0.193     0.024     0.167     0.407     0.767    18.879
```


And, as a table:

| # Branch         | avg   | min   | p50   | p95   | p99   | max    |
|------------------|-------|-------|-------|-------|-------|--------|
| ISPN-16329 (GET) | 0.192 | 0.016 | 0.167 | 0.383 | 0.727 | 20.911 |
| main (GET)       | 0.193 | 0.024 | 0.167 | 0.407 | 0.767 | 18.879 |
| ISPN-16329 (SET) | 0.633 | 0.120 | 0.599 | 0.791 | 1.415 | 16.463 |
| main (SET)       | 0.626 | 0.448 | 0.599 | 0.775 | 1.343 | 11.759 |