@jabolina 2024-08-29
Another big thing. I should probably create a unit test for each serializer. We didn't have a way to do that before. But I'll keep the PR open to gather suggestions.