@jabolina 2024-09-12
Updated with the suggestions and trying to tackle the points.

1. I've created the static `BiConsumer` on the `Resp3Response` class. It is not very good-looking, but it should avoid the allocations.

The second point was a bit more tricky. I've added an explicit sub-type for serializer that deals with nested objects so they can receive a `SerializationHint` that redirects to the appropriate method. I really didn't want to open the can of deeply nested objects or homogeneous elements. This means the serialization hints only work if the structure (list, set, map) contains elements of the same type. For example, a `List<byte[]>` we can hint it is an array of bulk strings. But for lists (`List.of(byte[], Map<String, byte[]>)`) that mix types, it will fall back to iterate until finding the correct serializer.