@wburns 2024-09-26
It appears some tests are timing out with the changes.

```
2024-09-26T18:49:12.3773350Z [TestSuiteProgress] Test failed: org.infinispan.server.resp.RespSingleNodeTest.destroy
2024-09-26T18:49:12.4789165Z Dumping threads of process 264061 to /home/runner/work/infinispan/infinispan/server/resp/threaddump-RespSingleNodeTest-20240926-1849-264061.log
2024-09-26T18:49:12.6984392Z Interrupted thread testng-RespSingleNodeTest (32).
2024-09-26T18:49:13.6234258Z Killed processes 264061
2024-09-26T18:49:13.9741939Z [INFO] 
2024-09-26T18:49:13.9742393Z [INFO] Results:
2024-09-26T18:49:13.9742981Z [INFO] 
2024-09-26T18:49:13.9743384Z [ERROR] Errors: 
2024-09-26T18:49:13.9745042Z [ERROR]   RespSingleNodeTest.testRenameWithEx:1115 » RedisCommandTimeout Command timed out after 15 second(s)
2024-09-26T18:49:13.9749689Z [ERROR]   RespSingleNodeTest.testRenamenx:1125 » RedisCommandTimeout Command timed out after 15 second(s)
2024-09-26T18:49:13.9751456Z [ERROR]   RespSingleNodeTest.testScanCount:494 » RedisCommandTimeout Command timed out after 15 second(s)
2024-09-26T18:49:13.9753178Z [ERROR]   RespSingleNodeTest.testScan:475 » RedisCommandTimeout Command timed out after 15 second(s)
```