@jabolina 2024-09-30
The blocking on CI is caused by [ISPN-16699](https://issues.redhat.com/browse/ISPN-16699) (#12947). I'll rebase once that's done.